Fixed segment header offset which was incorrectly reduced to 16bit

+++ includes/acl/compression/stream/write_segment_data.h
@@ -36,7 +36,7 @@
 
 namespace acl
 {
-	inline void write_segment_headers(const ClipContext& clip_context, const CompressionSettings& settings, SegmentHeader* segment_headers, uint16_t segment_headers_start_offset)
+	inline void write_segment_headers(const ClipContext& clip_context, const CompressionSettings& settings, SegmentHeader* segment_headers, uint32_t segment_headers_start_offset)
 	{
 		const uint32_t format_per_track_data_size = get_format_per_track_data_size(clip_context, settings.rotation_format, settings.translation_format, settings.scale_format);
